requirements.txt
flask
python-dotenv
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
openai
flask_cors
pillow
bs4
gradio
sqlalchemy
alembic
psycopg2-binary